+++ src/components/CrateTabs/CrateTabs.tsx
@@ -7,7 +7,7 @@ import { QUERY_PARAM_KEY_ACTIVE_TAB } from 'constants/defaults';
 export type CrateTabsProps = TabsProps & {
   items: NonNullable<TabsProps['items']>;
   defaultActiveKey: NonNullable<TabsProps['defaultActiveKey']>;
-  queryParamKeyActiveTab?: string;
+  queryParamKeyActiveTab?: string | null;
   indentTabBar?: boolean; // indent the tabs (but not the content) for situations where the tabs are pushed up against a container
 };
 
@@ -23,9 +23,9 @@ function CrateTabs({
   const navigate = useNavigate();
   // preselected tabs set via URL search params will override
   // the default active key
-  const searchParamValue = new URLSearchParams(location.search).get(
-    queryParamKeyActiveTab,
-  );
+  const searchParamValue = queryParamKeyActiveTab
+    ? new URLSearchParams(location.search).get(queryParamKeyActiveTab)
+    : null;
   const childKeysIncludesSearchParam = searchParamValue
     ? items
         .filter(({ disabled = false }) => !disabled)
@@ -53,9 +53,11 @@ function CrateTabs({
   // URL bar from showing a stale search param when the tab changes and
   // means the browser back button can be used.
   const handleTabClick = (key: string) => {
-    const searchParams = new URLSearchParams(location.search);
-    searchParams.set(queryParamKeyActiveTab, key);
-    navigate(`?${searchParams.toString()}`);
+    if (queryParamKeyActiveTab) {
+      const searchParams = new URLSearchParams(location.search);
+      searchParams.set(queryParamKeyActiveTab, key);
+      navigate(`?${searchParams.toString()}`);
+    }
   };
 
   if (indentTabBar) {