+++ src/components/SQLEditor/annotationUtils.ts
@@ -1,5 +1,5 @@
 import { Ace } from 'ace-builds';
-import { QueryResults } from 'types/query';
+import { QueryResultError, QueryResults } from 'types/query';
 
 export const annotate = (
   ace: Ace.Editor,
@@ -17,33 +17,47 @@ export const annotate = (
   }
 
   if (Array.isArray(results)) {
-    const err = results.find(r => r.error);
+    const err = results.find(r => 'error' in r) as QueryResultError;
     if (!err) {
       return;
     }
-    const q = err.original_query;
-    if (!q) {
+
+    if (typeof err.line !== 'undefined') {
+      return [
+        {
+          row: err.line,
+          column: 0,
+          text: 'error' in err ? err.error.message : '',
+          type: 'error',
+        },
+      ];
+    }
+
+    if (!err.original_query) {
       return;
     }
+    const query = err.original_query.query;
     // we take the 1st 100 lines and see if any of them are contained in the
     // query that failed. A rather naive approach.
     const lines = ace.getSession().getLines(0, 100);
-    const trimmed = q.trim();
-    const pos = lines
-      .map((v, i) => {
-        const search = v.trim();
-        if (search.length > 0 && trimmed.startsWith(search)) {
-          return i;
-        }
-        return -1;
-      })
-      .find(v => v > 0);
-    if (pos) {
+
+    let currentString = '';
+    let pos: number | null = null;
+    let lineIndex = lines.length - 1;
+    for (const line of lines.reverse()) {
+      currentString = `${line}${lineIndex < lines.length ? '\n' : ''}${currentString}`;
+      if (currentString.startsWith(query)) {
+        pos = lineIndex;
+      }
+      lineIndex--;
+    }
+
+    if (pos !== null) {
       return [
         {
           row: pos,
           column: 0,
-          text: err.error?.message || '',
+          text: 'error' in err ? err.error.message : '',
           type: 'error',
         },
       ];
@@ -51,7 +65,7 @@ export const annotate = (
     return;
   }
 
-  if (results.error) {
+  if ('error' in results) {
     // We always show the 1st line and 1st column as the source of the error.
     return [
       {